backend/requirements.txt
django==4.2
pillow>=9.5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
python-dotenv==1.0.0
psycopg2-binary>=2.9.6
celery[redis]>=5.0.5
redis==4.6.0
gunicorn==20.1.0
django-celery-beat
django-celery-results
python-decouple==3.8
django-allauth~=0.54.0
PyJWT~=2.6.0
dj-rest-auth==4.0.0
django-filter==23.2
djangorestframework-simplejwt==5.3.0
django-countries==7.5.1
stripe==6.5.0
whitenoise==6.5.0
django-redis==5.4.0
drf-yasg==1.21.7